CI: try macOS and Linux builds (#1)
* CI: try macOS and Linux builds (artifacts only); fix POSIX CMake

- Split the icon resources: TsFileViewer.rc compiles only on Windows,
  app.qrc (window icon) on every platform
- POSIX link list gains Qt5::Concurrent and Qt5::Network (both used by
  the app regardless of platform)
- New posix matrix job (ubuntu-24.04, macos-13 x86_64) building with
  Ninja and uploading tar.gz artifacts; no runtime bundling and no
  release upload yet - Windows stays the shipped platform

* CI: gcc_64 arch for Qt 5 Linux; macos-15-intel; sanitize PR ref names

- Qt 5.15.2 Linux binaries live under the gcc_64 arch (linux64 is Qt 6)
- macos-13 queue is dead (image retired); macos-15-intel is the last
  x86_64 image matching Qt 5.15.2's macos arch
- PR ref_name is 'N/merge': '/' breaks zip/tar paths in both the
  Windows stage step and the POSIX package step

* CI: clang_64 is the Qt 5 macOS arch name (macos is Qt 6)

* Fix qint64/int64_t template mismatches (long long vs long on Linux)

qMin/qMax/std::min/std::max need identical argument types; MSVC maps
both qint64 and int64_t to long long so this only surfaces on gcc/clang

* Link libz.a on POSIX: zlib's CMake renames zlibstatic to z on UNIX

---------

+++ Tools/SpanProbe.cpp
@@ -90,8 +90,12 @@ int main(int argc, char** argv)
                             }
                             else
                             {
-                                firstTs = qMin(firstTs, st->start_time_);
-                                lastTs = qMax(lastTs, st->get_end_time());
+                                // qint64 vs int64_t mismatch long long/long
+                                // on Linux: cast to the common qint64.
+                                firstTs = qMin(firstTs,
+                                               qint64(st->start_time_));
+                                lastTs = qMax(lastTs,
+                                              qint64(st->get_end_time()));
                             }
                         }
                     }
